converted the SEO module to use the action controllers
