+++ ACP3/Modules/ACP3/Seo/Controller/Admin/Index/Index.php
@@ -1,0 +1,103 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\Seo\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Seo;
+
+/**
+ * Class Index
+ * @package ACP3\Modules\ACP3\Seo\Controller\Admin\Index
+ */
+class Index extends Core\Modules\AdminController
+{
+    /**
+     * @var \ACP3\Modules\ACP3\Seo\Model\SeoRepository
+     */
+    protected $seoRepository;
+
+    /**
+     * Index constructor.
+     *
+     * @param \ACP3\Core\Modules\Controller\AdminContext $context
+     * @param \ACP3\Modules\ACP3\Seo\Model\SeoRepository $seoRepository
+     */
+    public function __construct(
+        Core\Modules\Controller\AdminContext $context,
+        Seo\Model\SeoRepository $seoRepository
+    ) {
+        parent::__construct($context);
+
+        $this->seoRepository = $seoRepository;
+    }
+
+    /**
+     * @return array
+     */
+    public function execute()
+    {
+        $seo = $this->seoRepository->getAllInAcp();
+
+        /** @var Core\Helpers\DataGrid $dataGrid */
+        $dataGrid = $this->get('core.helpers.data_grid');
+        $dataGrid
+            ->setResults($seo)
+            ->setRecordsPerPage($this->user->getEntriesPerPage())
+            ->setIdentifier('#acp-table')
+            ->setResourcePathDelete('admin/seo/index/delete')
+            ->setResourcePathEdit('admin/seo/index/edit');
+
+        $dataGrid
+            ->addColumn([
+                'label' => $this->translator->t('seo', 'uri'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\TextColumnRenderer::class,
+                'fields' => ['uri'],
+                'default_sort' => true
+            ], 60)
+            ->addColumn([
+                'label' => $this->translator->t('seo', 'alias'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\TextColumnRenderer::class,
+                'fields' => ['alias'],
+            ], 50)
+            ->addColumn([
+                'label' => $this->translator->t('seo', 'keywords'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\TextColumnRenderer::class,
+                'fields' => ['keywords'],
+            ], 40)
+            ->addColumn([
+                'label' => $this->translator->t('seo', 'description'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\TextColumnRenderer::class,
+                'fields' => ['description'],
+            ], 30)
+            ->addColumn([
+                'label' => $this->translator->t('seo', 'robots'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\ReplaceValueColumnRenderer::class,
+                'fields' => ['robots'],
+                'custom' => [
+                    'search' => [0, 1, 2, 3, 4],
+                    'replace' => [
+                        $this->translator->t('seo', 'robots_use_system_default',
+                            ['%default%' => $this->seo->getRobotsSetting()]),
+                        $this->translator->t('seo', 'robots_index_follow'),
+                        $this->translator->t('seo', 'robots_index_nofollow'),
+                        $this->translator->t('seo', 'robots_noindex_follow'),
+                        $this->translator->t('seo', 'robots_noindex_nofollow')
+                    ]
+                ]
+            ], 20)
+            ->addColumn([
+                'label' => $this->translator->t('system', 'id'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\IntegerColumnRenderer::class,
+                'fields' => ['id'],
+                'primary' => true
+            ], 10);
+
+        return [
+            'grid' => $dataGrid->render(),
+            'show_mass_delete_button' => count($seo) > 0
+        ];
+    }
+}
