+++ ACP3/Modules/ACP3/Seo/Resources/config/controllers.yml
@@ -1,0 +1,44 @@
+parameters:
+  seo.controller.admin.index.create.fqdn: 'ACP3\Modules\ACP3\Seo\Controller\Admin\Index\Create'
+  seo.controller.admin.index.delete.fqdn: 'ACP3\Modules\ACP3\Seo\Controller\Admin\Index\Delete'
+  seo.controller.admin.index.edit.fqdn: 'ACP3\Modules\ACP3\Seo\Controller\Admin\Index\Edit'
+  seo.controller.admin.index.index.fqdn: 'ACP3\Modules\ACP3\Seo\Controller\Admin\Index\Index'
+  seo.controller.admin.index.settings.fqdn: 'ACP3\Modules\ACP3\Seo\Controller\Admin\Index\Settings'
+
+services:
+  seo.controller.admin.index.create:
+    class: %seo.controller.admin.index.create.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.helpers.formToken
+      - @seo.model.seorepository
+      - @seo.validator
+
+  seo.controller.admin.index.delete:
+    class: %seo.controller.admin.index.delete.fqdn%
+    arguments:
+      - @core.context.admin
+      - @seo.cache
+      - @seo.model.seorepository
+
+  seo.controller.admin.index.edit:
+    class: %seo.controller.admin.index.edit.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.helpers.formToken
+      - @seo.cache
+      - @seo.model.seorepository
+      - @seo.validator
+
+  seo.controller.admin.index.index:
+    class: %seo.controller.admin.index.index.fqdn%
+    arguments:
+      - @core.context.admin
+      - @seo.model.seorepository
+
+  seo.controller.admin.index.settings:
+    class: %seo.controller.admin.index.settings.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.helpers.formToken
+      - @seo.validation.admin_settings_form_validation
